pyproject.toml
[project]
name = "hackudc-2026"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "gdown>=5.2.1",
    "ipykernel>=7.2.0",
    "ipywidgets>=8.1.8",
    "onnxruntime>=1.24.2",
    "pandas>=3.0.1",
    "pillow>=12.1.1",
    "rembg>=2.0.72",
    "requests>=2.32.5",
    "setuptools<70",
    "torch>=2.10.0",
    "tqdm>=4.67.3",
    "transformers>=5.2.0",
    "ultralytics==8.0.21",
    "ultralyticsplus==0.0.23",
]
